+++ packages/media/media-picker/src/components/dropzone/__tests__/unit/index.spec.ts
@@ -27,7 +27,6 @@ describe('Dropzone', () => {
       collection: '',
     },
     container,
-    useNewUploadService: true,
   };
   // Helper functions
   const createDragOverOrDropEvent = (