+++ packages/editor/editor-common/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/editor-common
 
+## 14.0.12
+- [patch] ED-5172 pressing enter after media single in list no longer deletes list [74824f8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/74824f8)
+
 ## 14.0.11
 - [patch] Updated dependencies [b12f7e6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b12f7e6)
   - @atlaskit/profilecard@4.0.8